requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
python-dotenv
jinja2
python-multipart